# Fix performance of class concatenation in `new_data_frame()`