--- a/R/utils.R
+++ b/R/utils.R
@@ -212,3 +212,7 @@ import_from <- function(ns, names, env = caller_env()) {
   objs <- env_get_list(ns_env(ns), names)
   env_bind(env, !!!objs)
 }
+
+fast_c <- function(x, y) {
+  .Call(vctrs_fast_c, x, y)
+}