--- a/src/init.c
+++ b/src/init.c
@@ -121,6 +121,7 @@ extern SEXP vctrs_s3_find_method(SEXP, SEXP, SEXP);
 extern SEXP vctrs_implements_ptype2(SEXP);
 extern SEXP vctrs_ptype2_dispatch_native(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP vctrs_cast_dispatch_native(SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP vctrs_fast_c(SEXP, SEXP);
 
 
 // Maturing
@@ -259,6 +260,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_implements_ptype2",          (DL_FUNC) &vctrs_implements_ptype2, 1},
   {"vctrs_ptype2_dispatch_native",     (DL_FUNC) &vctrs_ptype2_dispatch_native, 5},
   {"vctrs_cast_dispatch_native",       (DL_FUNC) &vctrs_cast_dispatch_native, 5},
+  {"vctrs_fast_c",                     (DL_FUNC) &vctrs_fast_c, 2},
   {NULL, NULL, 0}
 };