--- a/src/vctrs.h
+++ b/src/vctrs.h
@@ -17,6 +17,10 @@ extern bool vctrs_debug_verbose;
 
 
 typedef R_xlen_t r_ssize;
+#define R_SSIZE_MAX R_XLEN_T_MAX
+
+#define r_length Rf_length
+#define r_new_vector Rf_allocVector
 
 // An ERR indicates either a C NULL in case of no error, or a
 // condition object otherwise