Closes #1174.

```r
sc <- c("tbl_df", "tbl")
c <- c("tbl_df", "tbl", "data.frame")

bench::mark(
  new_data_frame(class = sc),
  new_data_frame(class = c),
  new_data_frame(),
  iterations = 1e6,
  check = FALSE
)
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`  n_itr  n_gc
#>   <bch:expr>                 <bch> <bch:tm>     <dbl> <bch:byt>    <dbl>  <int> <dbl>
#> 1 new_data_frame(class = sc) 834ns   1.06µs   824160.        0B     33.8 999959    41
#> 2 new_data_frame(class = c)  834ns   1.05µs   836700.        0B     34.3 999959    41
#> 3 new_data_frame()           719ns    881ns  1004085.        0B     35.1 999965    35
```