@lionel- 2020-07-06
Also start implementing safe routines for arithmetic of signed integer types.